refactor(product-agent): IA 重构 P0 — 两级结构 + 新建独立页 + tab 调整
按用户结构调整需求重构信息架构:
- 拆分:/product-agent 改为产品管理列表页(卡片网格 + 新建/修改/删除/筛选),
  点产品进 /product-agent/p/:id 单产品视图
- SingleProductView:读 :productId,header(返回列表+删除)+ 八 tab
- 需求/功能「新建」改为独立页面 /product-agent/p/:productId/:kind/new
  (ProductObjectDetailPage 新增 new 模式 + CreateObjectForm)
- 单产品 tab 调序:缺陷移到客户之前
- 升级申请从顶层 tab 并入「版本」tab
- 详情/新建路由统一到 /product-agent/p/:productId/:kind/:id
- 删除旧 ProductAgentPage(列表+详情合一)

纯前端重构,复用既有端点,无后端改动。
自测:tsc 全绿 + lint 零告警 + navCoverage 通过(参数化子路由自动豁免)
