+++ changelogs/2026-06-02_product-agent-ia-p0.md
@@ -1,0 +1,1 @@
+| refactor | prd-admin | 产品管理智能体 IA 重构 P0：拆分为「产品管理列表页」(/product-agent,增删改查+筛选)与「单产品视图」(/product-agent/p/:id);需求/功能新建改为独立页面;单产品 tab 调序(缺陷移到客户前);升级申请并入版本 tab |
