+++ prd-admin/src/pages/product-agent/index.ts
@@ -1,2 +1,3 @@
-export { ProductAgentPage } from './ProductAgentPage';
+export { ProductsListPage } from './ProductsListPage';
+export { SingleProductView } from './SingleProductView';
 export { ProductObjectDetailPage } from './ProductObjectDetailPage';
